requirements.txt
et-xmlfile==1.1.0
openpyxl==3.1.4
psycopg==3.1.19
python-dotenv==1.0.1
typing_extensions==4.12.2
